+++ src/Common/Log.hpp
@@ -52,7 +52,7 @@ namespace HighQueue
         };
 
         static Level runtimeLevel;
-        static void log(Level level, const char * file, const char * function, uint16_t line, const std::string & message);
+        static void log(Level level, const char * file, const char * function, long line, const std::string & message);
         static bool isEnabled(Level level);
         static Level getLevel();
         static void setLevel(Level level);
